Refine thresholding and merge top-k accuracy
